Move 'Failed logs' fixture in order to allow easy composition
